JETTY-629 do not hold timeout lock during expiry

+++ modules/util/src/main/java/org/mortbay/thread/Timeout.java
@@ -26,21 +26,28 @@ import org.mortbay.log.Log;
  * <p>
  * The nested class Task should be extended by users of this class to obtain call back notification of 
  * expiries. 
- * <p>
- * This class is not synchronized and the caller must protect against multiple thread access.
  * 
  * @author gregw
  *
  */
 public class Timeout
 {
-    
+    private Object _lock;
     private long _duration;
     private long _now=System.currentTimeMillis();
     private Task _head=new Task();
-    
+
+    /* ------------------------------------------------------------ */
     public Timeout()
     {
+        _lock=new Object();
+        _head._timeout=this;
+    }
+
+    /* ------------------------------------------------------------ */
+    public Timeout(Object lock)
+    {
+        _lock=lock;
         _head._timeout=this;
     }
 
@@ -63,21 +70,31 @@ public class Timeout
     }
 
     /* ------------------------------------------------------------ */
-    public void setNow()
+    public long setNow()
     {
-        _now=System.currentTimeMillis();
+        synchronized (_lock)
+        {
+            _now=System.currentTimeMillis();
+            return _now; 
+        }
     }
     
     /* ------------------------------------------------------------ */
     public long getNow()
     {
-        return _now;
+        synchronized (_lock)
+        {
+            return _now;
+        }
     }
 
     /* ------------------------------------------------------------ */
     public void setNow(long now)
     {
-        _now=now;
+        synchronized (_lock)
+        {
+            _now=now;
+        }
     }
 
     /* ------------------------------------------------------------ */
@@ -90,44 +107,64 @@ public class Timeout
      */
     public Task expired()
     {
-        long _expiry = _now-_duration;
-        
-        if (_head._next!=_head)
+        synchronized (_lock)
         {
-            Task task = _head._next;
-            if (task._timestamp>_expiry)
-                return null;
-            
-            synchronized (task)
+            long _expiry = _now-_duration;
+
+            if (_head._next!=_head)
             {
+                Task task = _head._next;
+                if (task._timestamp>_expiry)
+                    return null;
+
                 task.unlink();
                 task._expired=true;
+                return task;
             }
-            return task;
+            return null;
         }
-        return null;
     }
 
     /* ------------------------------------------------------------ */
-    public void tick()
+    public void tick(long now)
     {
-        long _expiry = _now-_duration;
-        
-        while (_head._next!=_head)
+        long _expiry = -1;
+
+        Task task=null;
+        while (true)
         {
-            Task task = _head._next;
-            if (task._timestamp>_expiry)
-                break;
-            
             try
             {
-                task.doExpire();
+                synchronized (_lock)
+                {
+                    if (_expiry==-1)
+                    {
+                        if (now!=-1)
+                            _now=now;
+                        _expiry = _now-_duration;
+                    }
+                        
+                    task= _head._next;
+                    if (task==_head || task._timestamp>_expiry)
+                        break;
+                    task.unlink();
+                    task._expired=true;
+                    task.expire();
+                }
+                
+                task.expired();
             }
             catch(Throwable th)
             {
                 Log.warn(Log.EXCEPTION,th);
             }
         }
+    }
+
+    /* ------------------------------------------------------------ */
+    public void tick()
+    {
+        tick(-1);
     }
 
     /* ------------------------------------------------------------ */
@@ -139,45 +176,58 @@ public class Timeout
     /* ------------------------------------------------------------ */
     public void schedule(Task task,long delay)
     {
-        if (task._timestamp!=0)
+        synchronized (_lock)
         {
-            task.unlink();
-            task._timestamp=0;
+            if (task._timestamp!=0)
+            {
+                task.unlink();
+                task._timestamp=0;
+            }
+            task._timeout=this;
+            task._expired=false;
+            task._delay=delay;
+            task._timestamp = delay==0?(_now+delay):(_now-_duration+delay);
+
+            Task last=_head._prev;
+            while (last!=_head)
+            {
+                if (last._timestamp <= task._timestamp)
+                    break;
+                last=last._prev;
+            }
+            last.link(task);
         }
-        task._expired=false;
-        task._delay=delay;
-        task._timestamp = delay==0?(_now+delay):(_now-_duration+delay);
-        
-        Task last=_head._prev;
-        while (last!=_head)
-        {
-            if (last._timestamp <= task._timestamp)
-                break;
-            last=last._prev;
-        }
-        last.setNext(task);
     }
 
 
     /* ------------------------------------------------------------ */
     public void cancelAll()
     {
-        _head._next=_head._prev=_head;
+        synchronized (_lock)
+        {
+            _head._next=_head._prev=_head;
+        }
     }
 
     /* ------------------------------------------------------------ */
     public boolean isEmpty()
     {
-        return _head._next==_head;
+        synchronized (_lock)
+        {
+            return _head._next==_head;
+        }
     }
 
     /* ------------------------------------------------------------ */
     public long getTimeToNext()
     {
-        if (_head._next==_head)
-            return -1;
-        long to_next = _duration+_head._next._timestamp-_now;
-        return to_next<0?0:to_next;
+        synchronized (_lock)
+        {
+            if (_head._next==_head)
+                return -1;
+            long to_next = _duration+_head._next._timestamp-_now;
+            return to_next<0?0:to_next;
+        }
     }
 
     /* ------------------------------------------------------------ */
@@ -218,16 +268,19 @@ public class Timeout
         long _timestamp=0;
         boolean _expired=false;
 
+        /* ------------------------------------------------------------ */
         public Task()
         {
             _next=_prev=this;
         }
 
+        /* ------------------------------------------------------------ */
         public long getTimestamp()
         {
             return _timestamp;
         }
-        
+
+        /* ------------------------------------------------------------ */
         public long getAge()
         {
             Timeout t = _timeout;
@@ -235,28 +288,24 @@ public class Timeout
                 return t._now-_timestamp;
             return 0;
         }
-        
-        public void unlink()
+
+        /* ------------------------------------------------------------ */
+        private void unlink()
         {
             _next._prev=_prev;
             _prev._next=_next;
             _next=_prev=this;
-            _timeout=null;
             _expired=false;
         }
 
-        public void setNext(Task task)
+        /* ------------------------------------------------------------ */
+        private void link(Task task)
         {
-            if (_timeout==null || 
-                task._timeout!=null && task._timeout!=_timeout ||    
-                task._next!=task)
-                throw new IllegalStateException();
             Task next_next = _next;
             _next._prev=task;
             _next=task;
             _next._next=next_next;
             _next._prev=this;   
-            _next._timeout=_timeout;
         }
         
         /* ------------------------------------------------------------ */
@@ -266,7 +315,6 @@ public class Timeout
          */
         public void schedule(Timeout timer)
         {
-            unlink();
             timer.schedule(this);
         }
         
@@ -277,20 +325,19 @@ public class Timeout
          */
         public void schedule(Timeout timer, long delay)
         {
-            unlink();
             timer.schedule(this,delay);
         }
         
         /* ------------------------------------------------------------ */
         /** Reschedule the task on the current timeout.
-         * The task timeout is rescheduled as if it had been canceled and
+         * The task timeout is rescheduled as if it had been cancelled and
          * scheduled on the current timeout.
          */
         public void reschedule()
         {
-            Timeout timer = _timeout;
-            unlink();
-            timer.schedule(this,_delay);
+            Timeout timeout = _timeout;
+            if (timeout!=null)
+                timeout.schedule(this,_delay);
         }
         
         /* ------------------------------------------------------------ */
@@ -300,7 +347,15 @@ public class Timeout
         public void cancel()
         {
             _timestamp=0;
-            unlink();
+            Timeout timeout = _timeout;
+            if (timeout!=null)
+            {
+                synchronized (timeout)
+                {
+                    _timestamp=0;
+                    unlink();
+                }
+            }
         }
         
         /* ------------------------------------------------------------ */
@@ -326,23 +381,6 @@ public class Timeout
          */
         public void expired(){}
 
-        /* ------------------------------------------------------------ */
-        private void doExpire()
-        {
-            try
-            {
-                synchronized (this)
-                {
-                    unlink();
-                    _expired=true;
-                    expire();
-                }
-            }
-            finally
-            {
-                expired();
-            }
-        }
     }
 
 }
